Refactor enums
