+++ frontend/app.py
@@ -1,21 +1,5 @@
 import streamlit as st
-from enum import Enum
-
-
-class Expertise(Enum):
-    INTRODUCTORY_CHEMISTRY = "Introductory Chemistry 🧪"
-
-
-class InputField(Enum):
-    QUESTION = "question"
-    OPTIONS = "option"
-    ANSWER = "answer"
-
-
-class OutputField(Enum):
-    GUIDE = "guide"
-    CITATIONS = "citations"
-
+from enums import Expertise, InputField, OutputField
 
 st.set_page_config(
     page_title="Tango AI",
